Code review: fix buffer overflow in session ID generation, improve PRNG seeding, fix test race, add missing include, harden config parsing
- Fix critical buffer overflow: generate_session_id() used char buf[21] but
  needs 22 bytes for 'sess_' + 16 hex chars + NUL terminator. buf[21] was
  writing past the end of the array (undefined behavior / stack corruption).
- Fix test race condition: ConfigManagerTest shared temp directory
  /tmp/cppsim_test across all parallel test instances; one test's TearDown
  could delete the directory while another was still using it. Now uses
  unique per-instance directories with random suffixes.
- Fix missing #include <string_view> in protocol.hpp: the header uses
  std::string_view in set_error_logger() but didn't include the header,
  relying on transitive includes (non-portable).
- Improve PRNG seeding: mt19937 was seeded from a single 32-bit
  random_device value, leaving most of the 624-word state deterministic.
  Now uses seed_seq populated with state_size (624) random words for full
  entropy initialization of session ID generation.
- Harden config JSON parsing: get<int>() calls on numeric fields could
  throw on unsigned JSON values. Changed to get<int64_t>() with appropriate
  casts for robustness across JSON number representations.

+++ src/server/runtime_config_manager.cpp
@@ -95,7 +95,7 @@ bool runtime_config_manager::load_from_json(const nlohmann::json& config_json) n
 
         // Load values from JSON with per-field clamping
         if (config_json.contains("max_connections") && config_json["max_connections"].is_number()) {
-            new_max_connections = config_json["max_connections"].get<int>();
+            new_max_connections = static_cast<int>(config_json["max_connections"].get<int64_t>());
             if (new_max_connections < 1 || new_max_connections > 10000) {
                 log_error("[RuntimeConfig] Invalid max_connections, using default");
                 new_max_connections = config::MAX_CONNECTIONS;
@@ -103,7 +103,7 @@ bool runtime_config_manager::load_from_json(const nlohmann::json& config_json) n
         }
         
         if (config_json.contains("handshake_timeout") && config_json["handshake_timeout"].is_number()) {
-            new_handshake_timeout = std::chrono::seconds(config_json["handshake_timeout"].get<int>());
+            new_handshake_timeout = std::chrono::seconds(config_json["handshake_timeout"].get<int64_t>());
             if (new_handshake_timeout < std::chrono::seconds(1) || new_handshake_timeout > std::chrono::seconds(300)) {
                 log_error("[RuntimeConfig] Invalid handshake_timeout, using default");
                 new_handshake_timeout = config::HANDSHAKE_TIMEOUT;
@@ -135,7 +135,7 @@ bool runtime_config_manager::load_from_json(const nlohmann::json& config_json) n
         }
         
         if (config_json.contains("rate_limit_window") && config_json["rate_limit_window"].is_number()) {
-            new_rate_limit_window = std::chrono::seconds(config_json["rate_limit_window"].get<int>());
+            new_rate_limit_window = std::chrono::seconds(config_json["rate_limit_window"].get<int64_t>());
             if (new_rate_limit_window < std::chrono::seconds(1) || new_rate_limit_window > std::chrono::seconds(60)) {
                 log_error("[RuntimeConfig] Invalid rate_limit_window, using default");
                 new_rate_limit_window = config::RATE_LIMIT_WINDOW;
@@ -143,7 +143,7 @@ bool runtime_config_manager::load_from_json(const nlohmann::json& config_json) n
         }
         
         if (config_json.contains("max_backoff") && config_json["max_backoff"].is_number()) {
-            new_max_backoff = std::chrono::seconds(config_json["max_backoff"].get<int>());
+            new_max_backoff = std::chrono::seconds(config_json["max_backoff"].get<int64_t>());
             if (new_max_backoff < std::chrono::seconds(1) || new_max_backoff > std::chrono::seconds(300)) {
                 log_error("[RuntimeConfig] Invalid max_backoff, using default");
                 new_max_backoff = config::MAX_BACKOFF;
@@ -151,7 +151,7 @@ bool runtime_config_manager::load_from_json(const nlohmann::json& config_json) n
         }
         
         if (config_json.contains("ws_idle_timeout") && config_json["ws_idle_timeout"].is_number()) {
-            new_ws_idle_timeout = std::chrono::seconds(config_json["ws_idle_timeout"].get<int>());
+            new_ws_idle_timeout = std::chrono::seconds(config_json["ws_idle_timeout"].get<int64_t>());
             if (new_ws_idle_timeout < std::chrono::seconds(1) || new_ws_idle_timeout > std::chrono::hours(48)) {
                 log_error("[RuntimeConfig] Invalid ws_idle_timeout, using default");
                 new_ws_idle_timeout = config::WS_IDLE_TIMEOUT;
